Towards unified controller
Initially the goal was to have two different MCs, one for the LED controller and one for the remote panel.
This is moving towards have one code that can run either as LED controller or remote panel.
The configuration screen has been modified to support this.
There is also a better logic to erase characters.